refactor: moved config to amp-access-scripts
refactor:
- moved config to amp-access-scripts
- moved Epic Element to EpicBlockComponent
- moved EpicContentType to type definition file
