+++ packages/frontend/index.d.ts
@@ -89,6 +89,12 @@ type ReaderRevenuePosition =
     | 'ampHeader'
     | 'ampFooter';
 
+// tslint:disable-next-line:prefer-array-literal
+type EpicContent = Array<{
+    heading: string;
+    paragraphs: string;
+    highlightedText: string;
+}>;
 interface NavType {
     pillars: PillarType[];
     otherLinks: MoreType;
